app/requirements.txt
mysql-connector
altgraph==0.10.2
astroid==1.6.2
backports-abc==0.5
backports.functools-lru-cache==1.5
bdist-mpkg==0.5.0
beautifulsoup4==4.6.0
cachetools==2.0.1
certifi==2018.1.18
chardet==3.0.4
click==6.7
configparser==3.5.0
enum34==1.1.6
Flask>=0.12.3
Flask-WTF==0.14.2
futures==3.1.1
google==2.0.1
google-api-core==1.1.0
google-auth==1.4.1
google-cloud-core==0.28.1
google-cloud-monitoring==0.28.1
googleapis-common-protos==1.5.3
idna==2.6
isort==4.3.4
itsdangerous==0.24
Jinja2==2.10
lazy-object-proxy==1.3.1
lz4==0.8.2
macholib==1.5.1
MarkupSafe==1.0
# matplotlib==1.3.1
mccabe==0.6.1
# mercurial==4.8
modulegraph==0.10.4
nose==1.3.7
numpy==1.15.4
prometheus-client==0.4.2
protobuf==3.5.2.post1
py2app==0.7.3
pyasn1==0.4.2
pyasn1-modules==0.2.1
pylint==1.8.3
# pyOpenSSL==0.13.1
pyparsing==2.0.1
python-dateutil==1.5
pytz==2018.3
requests>=2.20.0
rsa==3.4.2
scipy==1.1.0
singledispatch==3.4.0.3
six==1.11.0
tornado==5.0.2
urllib3==1.22
uWSGI==2.0.17.1
virtualenv==15.2.0
Werkzeug==0.15.3
wrapt==1.10.11
WTForms==2.1

